Extract TestResult
