+++ src/TestResults/index.ts
@@ -1,2 +1,3 @@
 export * from './TestReconciliationState'
+export * from './TestResult'
 export * from './TestResultProvider'
